requirements.txt
asgiref==3.8.1
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.10.0
rest-framework-simplejwt==0.0.2
sqlparse==0.5.2
tzdata==2024.2
